requirements.txt
imageio>=2.34.0
imageio-ffmpeg>=0.6.0
Pillow>=10.0.0
PyYAML>=6.0
requests>=2.31.0
tqdm>=4.66.0
